+++ R/fit.R
@@ -54,7 +54,9 @@
 #'        in argument `times`.
 #' @param times A integer vector listing the set of times in order.
 #'        If `times=NULL`, then it is filled in as the vector of integers
-#'        from the minimum to maximum value of `data$time`.
+#'        from the minimum to maximum value of `data$time`.  Alternatively,
+#'        it could be the minimum value of `data$time` through future years,
+#'        such that the model can forecast those future years.
 #' @param variable_column A character string indicating the column of `data`
 #'        listing the variable for each sample, from the set of times
 #'        in argument `variables`.
@@ -106,13 +108,15 @@
 #' @importFrom corpcor pseudoinverse
 #' @importFrom methods is as
 #' @importFrom fmesher fm_evaluator fm_mesh_2d fm_fem
-#' @importFrom stats .getXlevels gaussian lm model.frame model.matrix
+#' @importFrom stats .getXlevels gaussian lm model.frame model.matrix update
 #'   model.offset model.response na.omit nlminb optimHess pnorm rnorm terms
 #'   update.formula binomial poisson predict
 #' @importFrom TMB MakeADFun sdreport
 #' @importFrom checkmate assertClass assertDataFrame checkInteger checkNumeric assertNumeric
-#' @importFrom Matrix Cholesky solve Matrix diag
+#' @importFrom Matrix Cholesky solve Matrix diag t mat2triplet
 #' @importFrom abind abind
+#' @importFrom insight get_response get_data
+#' @importFrom cv GetResponse cv
 #'
 #' @seealso Details section of [make_dsem_ram()] for a summary of the math involved with constructing the DSEM, and \doi{10.1111/2041-210X.14289} for more background on math and inference
 #' @seealso \doi{10.48550/arXiv.2401.10193} for more details on how GAM, SEM, and DSEM components are combined from a statistical and software-user perspective
@@ -161,6 +165,10 @@
 #'                 spatial_domain = mesh,
 #'                 space_term = "n <-> n, sd_n" )
 #'
+#' # Run crossvalidation
+#' CV = cv::cv( out, k = 4 )
+#' CV
+#'
 #' @useDynLib tinyVAST, .registration = TRUE
 #' @export
 tinyVAST <-
@@ -192,6 +200,12 @@ function( formula,
   assertClass(control, "tinyVASTcontrol")
   assertDataFrame(data)
   if(inherits(data,"tbl")) stop("`data` must be a data.frame and cannot be a tibble")
+
+  # Input conflicts
+  matched_call = match.call()
+  if( isTRUE(as.character(matched_call$family) == "family") ){
+    stop("Naming argument `family` as `family` conflicts with function `cv::cv`, please use `family = Family` or other name")
+  }
 
   # Haven't tested for extra levels
   tmpdata = droplevels(data)
@@ -300,10 +314,10 @@ function( formula,
     df_ram = data.frame(output$ram)
     ram_gamma = df_ram[df_ram$heads==2,,drop=FALSE]
     total_variance_h = tapply( as.numeric(ifelse(is.na(ram_gamma$start), 1, ram_gamma$start)),
-            INDEX = ram_gamma$from, FUN=\(x)sum(abs(x)) )
+            INDEX = ram_gamma$from, FUN=function(x)sum(abs(x)) )
     ram_rho = df_ram[df_ram$heads==1,,drop=FALSE]
     total_effect_h = tapply( as.numeric(ifelse(is.na(ram_rho$start), 1, ram_rho$start)),
-            INDEX = ram_rho$from, FUN=\(x)sum(abs(x)) )
+            INDEX = ram_rho$from, FUN=function(x)sum(abs(x)) )
     if( any(total_variance_h==0) && control$gmrf_parameterization=="separable" ){
       stop("Must use gmrf_parameterization=`projection` for the spacetime_term RAM supplied")
     }
@@ -349,10 +363,10 @@ function( formula,
     df_ram = data.frame(output$ram)
     ram2 = df_ram[df_ram$heads == 2,,drop=FALSE]
     total_variance_h = tapply( as.numeric(ifelse( is.na(ram2$start), 1, ram2$start)),
-            INDEX = ram2$from, FUN=\(x)sum(abs(x)) )
+            INDEX = ram2$from, FUN=function(x)sum(abs(x)) )
     ram1 = df_ram[df_ram$heads == 1,,drop=FALSE]
     total_effect_h = tapply( as.numeric(ifelse(is.na(ram1$start), 1, ram1$start)),
-            INDEX = ram1$from, FUN=\(x)sum(abs(x)) )
+            INDEX = ram1$from, FUN=function(x)sum(abs(x)) )
     if( any(total_variance_h==0) && control$gmrf_parameterization=="separable" ){
       stop("Must use options$gmrf_parameterization=`projection` for the space_term RAM supplied")
     }
@@ -410,7 +424,7 @@ function( formula,
 
   #
   Aepsilon_zz = cbind(Atriplet$i, Atriplet$j, t_i[Atriplet$i], c_i[Atriplet$i])
-  which_Arows = which(apply( Aepsilon_zz, MARGIN=1, FUN=\(x) all(!is.na(x)) & any(x>0) ))
+  which_Arows = which(apply( Aepsilon_zz, MARGIN=1, FUN=function(x) all(!is.na(x)) & any(x>0) ))
   which_Arows = which_Arows[ which(Atriplet$x[which_Arows] > 0) ]
   if( (nrow(spacetime_term_ram$output$ram)==0) & (nrow(delta_spacetime_term_ram$output$ram)==0) ){
     which_Arows = numeric(0)
@@ -420,7 +434,7 @@ function( formula,
 
   #
   Aomega_zz = cbind(Atriplet$i, Atriplet$j, c_i[Atriplet$i])
-  which_Arows = which(apply( Aomega_zz, MARGIN=1, FUN=\(x) all(!is.na(x)) ))
+  which_Arows = which(apply( Aomega_zz, MARGIN=1, FUN=function(x) all(!is.na(x)) ))
   which_Arows = which_Arows[ which(Atriplet$x[which_Arows] > 0) ]
   if( (nrow(space_term_ram$output$ram)==0) & (nrow(delta_space_term_ram$output$ram)==0) ){
     which_Arows = numeric(0)
@@ -440,28 +454,49 @@ function( formula,
     offset_i = gam_setup$offset
 
     # Extract and combine penalization matrices
-    S_list = lapply( seq_along(gam_setup$smooth), \(x) gam_setup$smooth[[x]]$S[[1]] )
-    S_kk = Matrix::.bdiag(S_list)       # join S's in sparse matrix
-    Sdims = unlist(lapply(S_list,nrow)) # Find dimension of each S
-    if(is.null(Sdims)) Sdims = vector(length=0)
+    #S_list = lapply( seq_along(gam_setup$smooth), function(x) gam_setup$smooth[[x]]$S[[1]] )
+    #S_kk = Matrix::.bdiag(S_list)       # join S's in sparse matrix
+    #Sdims = unlist(lapply(S_list,nrow)) # Find dimension of each S
+    #if(is.null(Sdims)) Sdims = vector(length=0)
 
-    # Get covariates
-    not_allowed <- vapply(c("t2(", "te("), \(.x)
+    # Warnings and errors
+    not_allowed <- vapply( c("t2("), function(.x)
       length(grep(.x, x=gam_setup$term.names, fixed=TRUE)) > 0, FUN.VALUE = logical(1L)
     )
-    if (any(not_allowed)) {
-      stop("Found t2() or te() smoothers. These are not yet implemented.", call. = FALSE)
+    if(any(not_allowed)) {
+      stop("Found t2() smoothers. These are not yet implemented.", call. = FALSE)
     }
-    which_se = grep( pattern="s(", x=gam_setup$term.names, fixed=TRUE )
+    which_experimental <- vapply( c("ti(", "te("), function(.x)
+      length(grep(.x, x=gam_setup$term.names, fixed=TRUE)) > 0, FUN.VALUE = logical(1L)
+    )
+    if(any(which_experimental)) {
+      if(isTRUE(control$verbose)) message("Found ti() or te() smoothers. These are experimental.")
+    }
 
-    # Extract and add names
+    # Extract and combine penalization matrices by block
+    S_list = lapply( seq_along(gam_setup$smooth), function(x) gam_setup$smooth[[x]]$S )
+    S_kk = Matrix::.bdiag( lapply(S_list, Matrix::.bdiag) )
+    Sdims = unlist( lapply(S_list, FUN = function(List){nrow(List[[1]])}) )
+    Sblock = unlist( lapply(S_list, length) )
+    if( sum(Sdims * Sblock) != nrow(S_kk) ) stop("Check constructor for smooths, term `Sblock`")
+
+    # Identify fixed vs. random effects
+    search_names = c( "s(", "te(", "ti(", "t2(" )
+    which_se = sapply( X = search_names,
+                       FUN = function(x){ grep( pattern=x, x=gam_setup$term.names, fixed=TRUE ) } )
+    which_se = sort( unlist(which_se), decreasing = FALSE )
+    if( length(which_se) != sum(Sdims) ) stop("Check constructor for smooths, term `which_se`")
+
+    # Extract design matrices for fixed and random effects and add names
     colnames(gam_setup$X) = gam_setup$term.names
     X_ij = gam_setup$X[,setdiff(seq_len(ncol(gam_setup$X)),which_se),drop=FALSE]
     Z_ik = gam_setup$X[,which_se,drop=FALSE]
+    if(is.null(Sdims)) Sdims = numeric()
+    if(is.null(Sblock)) Sblock = numeric()
 
     #
-    out = list( "X_ij"=X_ij, "Z_ik"=Z_ik, "S_kk"=S_kk, "Sdims"=Sdims, "y_i"=y_i,
-                "offset_i"=offset_i, "gam_setup"=gam_setup )
+    out = list( "X_ij"=X_ij, "Z_ik"=Z_ik, "S_kk"=S_kk, "Sdims"=Sdims, "Sblock" = Sblock,
+                "y_i"=y_i, "offset_i"=offset_i, "gam_setup"=gam_setup )
     return(out)
   }
   gam_basis = build_gam_basis( formula )
@@ -484,8 +519,8 @@ function( formula,
 
     # Construct log_sigma based on family
     pad_length = function(x){if(length(x)==1) c(x,99L) else x}
-    remove_last = \(x) x[-length(x)]
-    Nsigma_e = sapply( family, FUN=\(x){
+    remove_last = function(x) x[-length(x)]
+    Nsigma_e = sapply( family, FUN=function(x){
                        switch( x$family[length(x$family)],
                          "gaussian" = 1,
                          "tweedie" = 2,
@@ -500,7 +535,7 @@ function( formula,
     Edims_ez = cbind( "start"=remove_last(cumsum(c(0,Nsigma_e))), "length"=Nsigma_e )
 
     #
-    family_code = t(rbind(sapply( family, FUN=\(x){
+    family_code = t(rbind(sapply( family, FUN=function(x){
                        pad_length(c("gaussian" = 0,
                          "tweedie" = 1,
                          "lognormal" = 2,
@@ -511,15 +546,15 @@ function( formula,
                          "nbinom1" = 6,
                          "nbinom2" = 7)[x$family])
                        } )))
-    link_code = t(rbind(sapply( family, FUN=\(x){
+    link_code = t(rbind(sapply( family, FUN=function(x){
                        pad_length(c("identity" = 0,
                          "log" = 1,
                          "logit" = 2,
                          "cloglog" = 3 )[x$link])
                        } )))
     components = apply( family_code, MARGIN=1,
-                         FUN=\(x)sum(x != 99L) )
-    poisson_link_delta = sapply( family, FUN=\(x){
+                         FUN=function(x)sum(x != 99L) )
+    poisson_link_delta = sapply( family, FUN=function(x){
                          as.integer(isTRUE(x$type == "poisson_link_delta"))} )
     out = list( "family_code" = cbind(family_code),
                 "link_code" = cbind(link_code),
@@ -554,9 +589,17 @@ function( formula,
   # All interactions among features should come here
   ##############
 
+  # Make options
+  spatial_options = c(
+    spatial_method_code,
+    ifelse( control$gmrf_parameterization=="separable", 0, 1),
+    ifelse( isFALSE(control$get_rsr), 0, 1),
+    ifelse( isFALSE(control$extra_reporting), 0, 1)
+  )
+
   # make dat
   tmb_data = list(
-    spatial_options = c(spatial_method_code, ifelse(control$gmrf_parameterization=="separable",0,1) ),
+    spatial_options = spatial_options,
     y_i = gam_basis$y_i,
     X_ij = gam_basis$X_ij,
     Z_ik = gam_basis$Z_ik,
@@ -576,8 +619,10 @@ function( formula,
     Edims_ez = distributions$Edims_ez,
     S_kk = gam_basis$S_kk,
     Sdims = gam_basis$Sdims,
+    Sblock = gam_basis$Sblock,
     S2_kk = delta_gam_basis$S_kk,
     S2dims = delta_gam_basis$Sdims,
+    S2block = delta_gam_basis$Sblock,
     Aepsilon_zz = Aepsilon_zz - 1,     # Index form, i, s, t
     Aepsilon_z = Aepsilon_z,
     Aomega_zz = Aomega_zz - 1,     # Index form, i, s, t
@@ -625,19 +670,19 @@ function( formula,
   # make params
   tmb_par = list(
     alpha_j = rep(0,ncol(tmb_data$X_ij)),  # Spline coefficients
-    gamma_k = rep(0,ncol(tmb_data$Z_ik)),  # Spline coefficients
+    gamma_k = rep(0,sum(tmb_data$Sdims)),  # Spline coefficients ... length is sum() <= ncol(tmb_data$Z_ik) ... not equal when using te/ti/t2
     beta_z = as.numeric(ifelse(spacetime_term_ram$param_type==1, 0.01, 1)),  # as.numeric(.) ensures class-numeric even for length=0 (when it would be class-logical), so matches output from obj$env$parList()
     theta_z = as.numeric(ifelse(space_term_ram$param_type==1, 0.01, 1)),
     nu_z = as.numeric(ifelse(time_term_ram$param_type==1, 0.01, 1)),
-    log_lambda = rep(0,length(tmb_data$Sdims)), #Log spline penalization coefficients
+    log_lambda = rep(0,sum(tmb_data$Sblock)), #Log spline penalization coefficients
     log_sigmaxi_l = rep(0,ncol(tmb_data$W_il)),
 
     alpha2_j = rep(0,ncol(tmb_data$X2_ij)),  # Spline coefficients
-    gamma2_k = rep(0,ncol(tmb_data$Z2_ik)),  # Spline coefficients
+    gamma2_k = rep(0,sum(tmb_data$S2dims)),  # Spline coefficients ... length is sum() <= ncol(tmb_data$Z_ik) ... not equal when using te/ti/t2
     beta2_z = as.numeric(ifelse(delta_spacetime_term_ram$param_type==1, 0.01, 1)),  # as.numeric(.) ensures class-numeric even for length=0 (when it would be class-logical), so matches output from obj$env$parList()
     theta2_z = as.numeric(ifelse(delta_space_term_ram$param_type==1, 0.01, 1)),
     nu2_z = as.numeric(ifelse(delta_time_term_ram$param_type==1, 0.01, 1)),
-    log_lambda2 = rep(0,length(tmb_data$S2dims)), #Log spline penalization coefficients
+    log_lambda2 = rep(0,sum(tmb_data$S2block)), #Log spline penalization coefficients
     log_sigmaxi2_l = rep(0,ncol(tmb_data$W2_il)),
 
     log_sigma = rep( 0, sum(distributions$Nsigma_e) ),
@@ -714,6 +759,8 @@ function( formula,
   if( !is.null(control$tmb_par) ){
     # Check shape but not numeric values, and give informative error
     attr(tmb_par,"check.passed") = attr(control$tmb_par,"check.passed")
+    # Compare dimensions by multiplying both by zero
+    warning("Supplying `control$tmb_par`:  use carefully as it may crash your terminal")
     if( isTRUE(all.equal(control$tmb_par, tmb_par, tolerance=Inf)) ){
       tmb_par = control$tmb_par
     }else{
@@ -741,6 +788,13 @@ function( formula,
       tmb_random = tmb_random
     )
     return(out)
+  }
+
+  # Debugging speedup ... if walking through manually, save and reload inputs in case of R crash
+  if( FALSE ){
+    setwd( R'(C:\Users\James.Thorson\Desktop\Temporary (can be deleted!))' )
+    save.image( "debugging.RData" )
+    #load.image( "debugging.RData" )
   }
 
   if( FALSE ){
@@ -840,7 +894,7 @@ function( formula,
     rep = obj$report(obj$env$last.par.best),
     sdrep = sdrep,
     tmb_inputs = list(tmb_data=tmb_data, tmb_par=tmb_par, tmb_map=tmb_map, tmb_random=tmb_random),
-    call = match.call(),
+    call = matched_call,
     run_time = Sys.time() - start_time,
     internal = internal
   )
@@ -897,6 +951,10 @@ function( formula,
 #' @param suppress_nlminb_warnings whether to suppress uniformative warnings
 #'        from \code{nlminb} arising when a function evaluation is NA, which
 #'        are then replaced with Inf and avoided during estimation
+#' @param get_rsr Experimental option, whether to report restricted spatial
+#'        regression (RSR) adjusted estimator for covariate responses
+#' @param extra_reporting Whether to report a much larger set of quantities via
+#'        \code{obj$env$report()}
 #'
 #' @return
 #' An object (list) of class `tinyVASTcontrol`, containing either default or
@@ -921,7 +979,9 @@ function( nlminb_loops = 1,
           getJointPrecision = FALSE,
           calculate_deviance_explained = TRUE,
           run_model = TRUE,
-          suppress_nlminb_warnings = TRUE ){
+          suppress_nlminb_warnings = TRUE,
+          get_rsr = FALSE,
+          extra_reporting = FALSE ){
 
   gmrf_parameterization = match.arg(gmrf_parameterization)
 
@@ -944,7 +1004,9 @@ function( nlminb_loops = 1,
     getJointPrecision = getJointPrecision,
     calculate_deviance_explained = calculate_deviance_explained,
     run_model = run_model,
-    suppress_nlminb_warnings = suppress_nlminb_warnings
+    suppress_nlminb_warnings = suppress_nlminb_warnings,
+    get_rsr = get_rsr,
+    extra_reporting = extra_reporting
   ), class = "tinyVASTcontrol" )
 }