+++ R/utility.R
@@ -299,8 +299,11 @@ function( object,
     # u_zr[-object$obj$env$random,1]
   }else{
     u_zr = object$obj$env$last.par.best %o% rep(1, n_samples)
-    MC = object$obj$env$MC( keep=TRUE, n=n_samples, antithetic=FALSE )
-    u_zr[object$obj$env$random,] = attr(MC, "samples")
+    Q = object$obj$env$spHess( random = TRUE, par = object$obj$env$last.par.best )
+    MC = rmvnorm_prec( mu=object$obj$env$last.par.best[object$obj$env$lrandom()], prec=Q, n.sims=n_samples, seed=seed)
+    #MC = object$obj$env$MC( keep=TRUE, n=n_samples, antithetic=FALSE )
+    #u_zr[object$obj$env$random,] = attr(MC, "samples")
+    u_zr[object$obj$env$lrandom(),] = MC
   }
 
   # Extract variable for each sample
@@ -373,7 +376,9 @@ function( x,
     control$trace = 0
     control$verbose = FALSE
     control$calculate_deviance_explained = FALSE
-  
+    control$tmb_par = NULL
+    control$tmb_map = NULL
+
   # 
   control_initial = control
     control_initial$nlminb_loops = 0
@@ -442,4 +447,183 @@ named_list <- function(...) {
   stats::setNames(L, nm)
 }
 
+
+
+#' @title Extract covariance
+#'
+#' @description Extract the covariance resulting from a specified path structure
+#' and estimated parameters for a SEM or DSEM term in tinyVAST
+#'
+#' @param object Output from \code{\link{tinyVAST}}
+#' @param what Which SEM or DSEM term to extract
+#' @param pred Extract the term \code{what} for which linear predictor
+#' @param n_times The number of times to include when calculating covariance for a DSEM
+#'        component, i.e., \code{time_term} or \code{spacetime_term}.  If missing, the default
+#'        is to use the one more than the maximum specified lag (e.g., \code{n_times=2}
+#'        by default when the maximum \code{lag=1})
+#'
+#' @importFrom dsem make_matrices
+#'
+#' @details
+#' tinyVAST constructs the covariance from specified path structure and estimated parameters
+#'
+#' @return
+#' The covariance matrix among variables
+#'
+#' @examples
+#' # Simulate settings
+#' set.seed(101)
+#' theta_xy = 0.4
+#' n_x = n_y = 10
+#' n_c = 3           # Number of species
+#' n_f = 1           # Number of factors
+#' rho = 0.8
+#' resid_sd = 0.5
+#'
+#' # Simulate GMRFs
+#' R_s = exp(-theta_xy * abs(outer(1:n_x, 1:n_y, FUN="-")) )
+#' R_ss = kronecker(X=R_s, Y=R_s)
+#' delta_fs = mvtnorm::rmvnorm(n_c, sigma=R_ss )
+#'
+#' # Simulate loadings for two factors
+#' L_cf = matrix( rnorm(n_c^2), nrow=n_c )
+#' L_cf[,seq(from=n_f+1, to=n_c)] = 0
+#' L_cf = L_cf + resid_sd * diag(n_c)
+#'
+#' # Simulate correlated densities
+#' d_cs = L_cf %*% delta_fs
+#'
+#' # Shape into longform data-frame and add error
+#' Data = data.frame( expand.grid(species=1:n_c, x=1:n_x, y=1:n_y),
+#'                    "var"="logn", "z"=exp(as.vector(d_cs)) )
+#' Data$n = rnorm( n=nrow(Data), mean=Data$z, sd=1 )
+#'
+#' # make mesh
+#' mesh = fmesher::fm_mesh_2d( Data[,c('x','y')] )
+#'
+#' # Specify factor model with two factors and additional independent variance with shared SD
+#' sem = "
+#'   # Loadings matrix
+#'   f1 -> 1, l1
+#'   f1 -> 2, l2
+#'   f1 -> 3, l3
+#'
+#'   # Factor variance = 1
+#'   f1 <-> f1, NA, 1
+#'
+#'   # Shared residual variance
+#'   1 <-> 1, sd, 1
+#'   2 <-> 2, sd, 1
+#'   3 <-> 3, sd, 1
+#' "
+#'
+#' # fit model
+#' out = tinyVAST( space_term = sem,
+#'            data = Data,
+#'            formula = n ~ 0 + factor(species),
+#'            spatial_domain = mesh,
+#'            variables = c( "f1", "f2", 1:n_c ),
+#'            space_columns = c("x","y"),
+#'            variable_column = "species",
+#'            time_column = "time",
+#'            distribution_column = "dist" )
+#'
+#' # Extract covariance among species and factors, where
+#' # estimated covariance is obtained by ignoring factors
+#' V = term_covariance( out, what = "space_term", pred = "one" )
+#'
+#' @export
+term_covariance <-
+function( object,
+          what = c("space_term", "time_term", "spacetime_term"),
+          pred = c("one", "two"),
+          n_times = NULL ){
+
+  pred = match.arg(pred)
+  what = match.arg(what)
+
+  # Extract stuff
+  variables = object$internal$variables
+  if(what=="space_term"){
+    if( pred == "one" ){
+      parname = "theta_z"
+      ram = object$internal$space_term_ram$output$ram
+    }else{
+      parname = "theta2_z"
+      ram = object$internal$delta_space_term_ram$output$ram
+    }
+    model = data.frame( "direction" = ram$heads,
+                        "lag" = 0,
+                        "first" = variables[ram$from],
+                        "second" = variables[ram$to],
+                        "start" = ram$start,
+                        "parameter" = ram$parameter )
+
+    if(is.null(n_times)){
+      times = 1
+    }else{
+      times = seq_len(n_times)
+    }
+    var_names = variables
+  }
+  if(what=="time_term"){
+    if( pred == "one" ){
+      parname = "nu_z"
+      model = as.data.frame(object$internal$time_term_ram$output$model)
+    }else{
+      parname = "nu2_z"
+      model = as.data.frame(object$internal$delta_time_term_ram$output$model)
+    }
+    if(is.null(n_times)){
+      times = seq_len( max(as.numeric(model$lag)+1) )
+    }else{
+      times = seq_len(n_times)
+    }
+    var_names = apply( expand.grid(paste0("_lag",times-1),variables),
+                       FUN=function(vec){paste0(vec[2],vec[1])}, MARGIN=1 )
+  }
+  if(what=="spacetime_term"){
+    if( pred == "one" ){
+      parname = "beta_z"
+      model = as.data.frame(object$internal$spacetime_term_ram$output$model)
+    }else{
+      parname = "beta2_z"
+      model = as.data.frame(object$internal$delta_spacetime_term_ram$output$model)
+    }
+    if(is.null(n_times)){
+      times = seq_len( max(as.numeric(model$lag)+1) )
+    }else{
+      times = seq_len(n_times)
+    }
+    var_names = apply( expand.grid(paste0("_lag",times-1),variables),
+                       FUN=function(vec){paste0(vec[2],vec[1])}, MARGIN=1 )
+  }
+
+  if( nrow(model) > 0 ){
+    # Unpack stuff
+    if(is.null(object$internal$parlist)){
+      object$internal$parlist = object$obj$env$parList()
+    }
+    # Extract path matrix
+    matrices = dsem::make_matrices(
+      beta_p = object$internal$parlist[[parname]],
+      model = model,
+      times = times,
+      variables = variables
+    )
+    P_kk = matrices$P_kk
+    G_kk = matrices$G_kk
+
+    # Calcualte covariance (not precision, in case G is not invertible)
+    IminusP_kk = Diagonal(n = nrow(P_kk)) - P_kk
+    #Q_kk = t(IminusP_kk) %*% solve(t(G_kk) %*% G_kk) %*% IminusP_kk
+    V_kk = solve(IminusP_kk) %*% t(G_kk) %*% G_kk %*% solve(t(IminusP_kk))
+    dimnames(V_kk) = list( var_names, var_names )
+  }else{
+    V_kk = "selected term not present in model"
+  }
+
+  # Return
+  return(V_kk)
+}