+++ tests/testthat/test-smooths.R
@@ -52,7 +52,32 @@ test_that("Basic splines work", {
   expect_gt(cor(p_s, p_v), 0.999)
 })
 
-test_that("2D splines work (or throw an error for now)", {
+test_that("Specify spline-order in s() works", {
+  #skip_on_cran()
+  #skip_on_ci()
+
+  set.seed(1)
+  dat <- mgcv::gamSim(1, n = 400, scale = 2)
+  newdata = data.frame( x2 = seq(-1, 1, by = 0.01) )
+
+  # Cubic spline ... penalize 2nd derivative, so extrapolate line beyond range
+  m_m <- mgcv::gam(y ~ s( x2, m = 2), data = dat, REML = FALSE)
+  p_m2 = predict( m_m, newdata )
+  m_v <- tinyVAST(formula = y ~ s(x2, m = 2), data = dat)
+  p_v2 <- predict( m_v, newdata )
+  expect_gt(cor(p_v2, p_m2), 0.999)
+  matplot( x = newdata$x2, y = cbind( p_m2, p_v2 ), type="l")
+
+  # Quadratic spline ... penalize 1st derivative, so extrapolate constant beyond range
+  m_m <- mgcv::gam(y ~ s( x2, m = 1), data = dat, REML = FALSE)
+  p_m1 = predict( m_m, newdata )
+  m_v <- tinyVAST(formula = y ~ s(x2, m = 1), data = dat)
+  p_v1 <- predict( m_v, newdata )
+  expect_gt(cor(p_v1, p_m1), 0.999)
+  matplot( x = newdata$x2, y = cbind( p_m1, p_v1 ), type="l")
+})
+
+test_that("t2 splines throw an error for now", {
   set.seed(1)
   dat <- mgcv::gamSim(1, n = 400, scale = 2)
   dat$fac <- fac <- as.factor(sample(1:20, 400, replace = TRUE))
@@ -62,7 +87,31 @@ test_that("2D splines work (or throw an error for now)", {
   m_m <- mgcv::gam(y ~ t2(x1, x2), data = dat)
   m_s <- sdmTMB::sdmTMB(y ~ t2(x1, x2), data = dat, spatial = "off")
   expect_error(m_v <- tinyVAST(formula = y ~ t2(x1, x2), data = dat), regexp = "t2")
-  expect_error(m_v <- tinyVAST(formula = y ~ te(x1, x2), data = dat), regexp = "te")
+})
+
+test_that("ti and te splines work", {
+  set.seed(1)
+  dat <- mgcv::gamSim(1, n = 400, scale = 2)
+  dat$fac <- fac <- as.factor(sample(1:20, 400, replace = TRUE))
+  dat$y <- dat$y + model.matrix(~ fac - 1) %*% rnorm(20) * .5
+  dat$y <- as.numeric(dat$y)
+
+  form = y ~ s(x1) + s(x2) + ti(x1, x2)
+  m_m <- mgcv::gam( form, data = dat, method="ML")
+  m_v <- tinyVAST( form, data = dat)
+  p_m <- predict(m_m)
+  p_v <- predict(m_v)
+  expect_gt(cor(p_v, p_m), 0.98)
+  cbind( "tinyVAST" = m_v$internal$parlist$log_lambda, "mgcv" = log(m_m$sp) )
+
+  #
+  form = y ~ te(x1, x2)
+  m_m <- mgcv::gam( form, data = dat, method="ML")
+  m_v <- tinyVAST( form, data = dat)
+  p_m <- predict(m_m)
+  p_v <- predict(m_v)
+  expect_gt(cor(p_v, p_m), 0.98)
+  cbind( "tinyVAST" = m_v$internal$parlist$log_lambda, "mgcv" = log(m_m$sp) )
 })
 
 # modified from sdmTMB tests:
@@ -180,21 +229,25 @@ test_that("A model with s(x, bs = 'cc') works", {
   expect_gt(stats::cor(p_v, p_m), 0.999)
 })
 
-# test_that("A model with s() by variables works", {
-#   set.seed(1)
-#   dat <- mgcv::gamSim(4)
-#   m_mgcv <- mgcv::gam(y ~ fac + s(x2, by = fac) + s(x0), data = dat)
-#   p_mgcv <- predict(m_mgcv)
-#
-#   m_s <- sdmTMB::sdmTMB(formula = y ~ fac + s(x2, by = fac) + s(x0), data = dat, spatial = "off")
-#
-#   m_v <- tinyVAST(formula = y ~ fac + s(x2, by = fac) + s(x0), data = dat)
-#   expect_s3_class(m_v, "tinyVAST")
-#
-#   p_m <- predict(m_mgcv)
-#   p_v <- predict(m_v)
-#   p_s <- predict(m_s)$est
-#
-#   plot(p_m, p_v);abline(a = 0, b = 1)
-#   expect_gt(cor(p_v, p_m), 0.9999)
-# })
+test_that("A model with s() by variables works", {
+  set.seed(1)
+  # For some reason, it doesn't work on CI but does locally
+  skip_on_ci()
+  skip_on_cran()
+
+  dat <- mgcv::gamSim(4)
+  m_mgcv <- mgcv::gam(y ~ fac + s(x2, by = fac) + s(x0), data = dat)
+  p_mgcv <- predict(m_mgcv)
+
+  m_s <- sdmTMB::sdmTMB(formula = y ~ fac + s(x2, by = fac) + s(x0), data = dat, spatial = "off")
+
+  m_v <- tinyVAST(formula = y ~ fac + s(x2, by = fac) + s(x0), data = dat)
+  expect_s3_class(m_v, "tinyVAST")
+
+  p_m <- predict(m_mgcv)
+  p_v <- predict(m_v)
+  p_s <- predict(m_s)$est
+
+  plot(p_m, p_v);abline(a = 0, b = 1)
+  expect_gt(cor(p_v, p_m), 0.99)
+})